backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg[binary]==3.2.4
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
python-dotenv==1.0.1
anthropic==0.39.0
eval_type_backport==0.3.1
alembic==1.14.0
cryptography==43.0.3
apscheduler==3.10.4
httpx==0.27.2
itsdangerous==2.2.0
sentry-sdk[fastapi]==2.18.0
pytest==8.3.3
</content>
</invoke>